# docs/swagger.yaml
basePath: /
definitions:
  dto.FollowingRequest:
    properties:
      following_id:
        type: integer
    type: object
  dto.JWT:
    properties:
      token:
        type: string
    type: object
  dto.LikeRequest:
    properties:
      post_id:
        type: integer
      user_id:
        type: integer
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: sosmed89@gmail.com
        type: string
      password:
        example: "12345678"
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/dto.JWT'
      message:
        example: Data retrieved successfully
        type: string
      status:
        example: Success
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: sosmed89@gmail.com
        type: string
      password:
        example: sosmed123
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterResponse:
    properties:
      message:
        example: Data retrieved successfully
        type: string
      status:
        example: Success
        type: string
      user_id:
        type: integer
    type: object
  dto.ResponseError:
    properties:
      errors:
        example: failed get data
        type: string
      message:
        example: Failed get data
        type: string
      status:
        example: Error
        type: string
    type: object
  dto.ResponseSuccess:
    properties:
      message:
        example: Data retrieved successfully
        type: string
      status:
        example: Success
        type: string
    type: object
host: 72.62.120.77:8081
info:
  contact: {}
  description: Backend Services For Coffee Morning
  title: Coffee Morning Services
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseError'
      summary: User login
      tags:
      - Auth
  /auth/logout:
    delete:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      summary: Register new user
      tags:
      - Auth
  /interaction/content:
    post:
      consumes:
      - multipart/form-data
      description: Create a new user account
      parameters:
      - description: Upload Content Image
        in: formData
        name: content_file
        required: true
        type: file
      - description: Caption
        in: formData
        name: caption
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Post new content
      tags:
      - Interaction
  /interaction/following:
    post:
      consumes:
      - application/json
      description: Follow a new user
      parameters:
      - description: Follow new user
        in: body
        name: follow
        required: true
        schema:
          $ref: '#/definitions/dto.FollowingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Follow new user
      tags:
      - Interaction
  /interaction/like:
    post:
      consumes:
      - application/json
      description: Like new post
      parameters:
      - description: Like new posts
        in: body
        name: follow
        required: true
        schema:
          $ref: '#/definitions/dto.LikeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Like new posts
      tags:
      - Interaction
  /user/:
    patch:
      consumes:
      - multipart/form-data
      parameters:
      - description: Full Name
        in: formData
        name: fullname
        type: string
      - description: Bio
        in: formData
        name: bio
        type: string
      - description: Update Avatar
        in: formData
        name: avatar_file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Update Profile
      tags:
      - user
  /user/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseSuccess'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - BearerAuth: []
      summary: Get User Profiles
      tags:
      - user
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by space and JWT Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"